refactor: ThemeRequest 클래스 dto 패키지로 이동 개선
