+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,6 +3,7 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Theme;
+import roomescape.dto.ThemeAddRequest;
 import roomescape.repository.ThemeDao;
 
 @Service
